Fix typo in chapter 1 where page author was prefixed with "by" twice.
In chapter 1, titled "Why software testing?", the following author text can be found:
"by by Mauricio Aniche". Commit 3c9769 changed the template (layout) to prefix
page authors with the word "by", instead of writing in separately per page.
It seems that removing the "by [author]" was forgotten for the first chapter.

This commit simply fixes the above by removing the "by" in the author name in
chapter 1.